backend/requirements.txt
pillow==11.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
Authlib==1.3.1
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
grpcio==1.67.1
grpcio-health-checking==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
urllib3==2.2.3
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.9.2
